+++ src/resolvers/generationHash.resolver.ts
@@ -28,6 +28,7 @@ export class GenerationHashResolver implements Resolver {
     /**
      * Resolves generationHash. If not provided by the user, this is asked to the node.
      * @param {CreateProfileOptions} options - Command options.
+     * @throws {ExpectedError}
      * @returns {Promise<string>}
      */
     async resolve(options: CreateProfileOptions): Promise<string> {