+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.20.2] - 30-May-2020
+
+**Milestone**: RC7 0.9.5.1
+
+### Fixed
+
+- ``transaction --mode`` property to announce transactions in batches
+- Standard console.log format #276
+
 ## [0.20.1] - 27-May-2020
 
 **Milestone**: RC7 0.9.5.1
@@ -13,6 +22,7 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 - ``transaction vrfkeylink`` command 
 - ``transaction nodekeylink`` command 
 - ``transaction votingkeylink`` command 
+- Multisig transaction announcement
 
 ### Changed